Pull code from git automatically.

+++ shell-scripts/update-test.sh
@@ -5,15 +5,22 @@ export PATH=$M2:$PATH
 
 cd ~/pratilipi
 
+
+while true
+do
+git remote update
+if [ $(git rev-parse HEAD) != $(git rev-parse origin) ]
+then
 git fetch
 git reset --hard origin/master
-
+date +"Last Updated on - %m-%d-%Y %r"
 mvn clean
 mvn install
 ant
-
 # Update devo-pratilipi/default
 mvn appengine:update -Dapp.id=devo-pratilipi -Dapp.module=default
-
 # Update prod-pratilipi/gamma
 mvn appengine:update -Dapp.id=prod-pratilipi -Dapp.module=gamma
+fi
+sleep 300
+done
